+++ examples/multiaz-node-group/main.tf
@@ -5,51 +5,25 @@ provider "aws" {
 #####
 # VPC and subnets
 #####
-module "vpc" {
-  source  = "terraform-aws-modules/vpc/aws"
-  version = "2.64.0"
+data "aws_vpc" "default" {
+  default = true
+}
 
-  name = "simple-vpc"
-
-  cidr = "10.0.0.0/16"
-
-  azs             = ["eu-west-1a", "eu-west-1b", "eu-west-1c"]
-  private_subnets = ["10.0.1.0/24", "10.0.2.0/24", "10.0.3.0/24"]
-  public_subnets  = ["10.0.101.0/24", "10.0.102.0/24", "10.0.103.0/24"]
-
-  private_subnet_tags = {
-    "kubernetes.io/role/internal-elb" = "1"
-  }
-
-  public_subnet_tags = {
-    "kubernetes.io/role/elb" = "1"
-  }
-
-  enable_dns_hostnames   = true
-  enable_dns_support     = true
-  enable_nat_gateway     = true
-  enable_vpn_gateway     = true
-  single_nat_gateway     = true
-  one_nat_gateway_per_az = false
-
-  tags = {
-    "kubernetes.io/cluster/eks" = "shared",
-    Environment                 = "test"
-  }
+data "aws_subnet_ids" "all" {
+  vpc_id = data.aws_vpc.default.id
 }
 
 #####
 # EKS Cluster
 #####
-
 resource "aws_eks_cluster" "cluster" {
   enabled_cluster_log_types = []
-  name                      = "eks"
+  name                      = "eks-module-test-cluster"
   role_arn                  = aws_iam_role.cluster.arn
-  version                   = "1.18"
+  version                   = "1.20"
 
   vpc_config {
-    subnet_ids              = flatten([module.vpc.public_subnets, module.vpc.private_subnets])
+    subnet_ids              = data.aws_subnet_ids.all.ids
     security_group_ids      = []
     endpoint_private_access = "true"
     endpoint_public_access  = "true"
@@ -132,9 +106,12 @@ module "eks-node-group-a" {
 
   create_iam_role = false
 
-  cluster_name  = aws_eks_cluster.cluster.id
-  node_role_arn = aws_iam_role.main.arn
-  subnet_ids    = [module.vpc.private_subnets[0]]
+  cluster_name = aws_eks_cluster.cluster.id
+
+  node_group_name_prefix = "eks-test-group-ab-"
+  node_role_arn          = aws_iam_role.main.arn
+
+  subnet_ids = [sort(data.aws_subnet_ids.all.ids)[0]]
 
   desired_size = 1
   min_size     = 1
@@ -144,7 +121,20 @@ module "eks-node-group-a" {
 
   ec2_ssh_key = "eks-test"
 
-  kubernetes_labels = {
+  taints = [
+    {
+      key    = "test-1"
+      value  = null
+      effect = "NO_SCHEDULE"
+    },
+    {
+      key    = "test-2"
+      value  = "value-test"
+      effect = "NO_EXECUTE"
+    }
+  ]
+
+  labels = {
     lifecycle = "OnDemand"
     az        = "eu-west-1a"
   }
@@ -159,9 +149,12 @@ module "eks-node-group-b" {
 
   create_iam_role = false
 
-  cluster_name  = aws_eks_cluster.cluster.id
+  cluster_name = aws_eks_cluster.cluster.id
+
+  node_group_name = "eks-test-group-b"
+
   node_role_arn = aws_iam_role.main.arn
-  subnet_ids    = [module.vpc.private_subnets[1]]
+  subnet_ids    = [sort(data.aws_subnet_ids.all.ids)[1]]
 
   desired_size = 1
   min_size     = 1
@@ -171,7 +164,7 @@ module "eks-node-group-b" {
 
   ec2_ssh_key = "eks-test"
 
-  kubernetes_labels = {
+  labels = {
     lifecycle = "OnDemand"
     az        = "eu-west-1b"
   }
@@ -188,7 +181,7 @@ module "eks-node-group-c" {
 
   cluster_name  = aws_eks_cluster.cluster.id
   node_role_arn = aws_iam_role.main.arn
-  subnet_ids    = [module.vpc.private_subnets[2]]
+  subnet_ids    = [sort(data.aws_subnet_ids.all.ids)[2]]
 
   desired_size = 1
   min_size     = 1
@@ -196,7 +189,7 @@ module "eks-node-group-c" {
 
   ec2_ssh_key = "eks-test"
 
-  kubernetes_labels = {
+  labels = {
     lifecycle = "OnDemand"
     az        = "eu-west-1c"
   }